Resource Drop System

+++ GD3B-Exam/Assets/Scripts/ScriptableObjects/ResourceManager.cs
@@ -3,5 +3,5 @@ using UnityEngine;
 public class ResourceManager : MonoBehaviour
 {
     public ResourceSO resourceSo;
-    public float resourceAmount = 1;
+    public int resourceAmount = 1;
 }
